ADD: ADD: ESTABLISHING THE VALIDATOR. sigma_i <- ARec(k_i, v', params).

+++ src/main/java/Utils.java
@@ -78,7 +78,7 @@ public class Utils {
         return (int) (flooredTerm % g);
     }
 
-    public static int ARec(double k2, int v, double q, double m, double g) {
+    public static int ARec(double k2, double v, double q, double m, double g) {
         // floor(m * (v / g - k_2 / q))
         long flooredTerm = (long) Math.floor(m * (v / g - k2 / q));
         return (int) (flooredTerm % m);
